move fan miot implementation to fan_miot.py
