+++ docs/api/miio.fan_common.rst
@@ -1,0 +1,7 @@
+miio.fan\_common module
+=======================
+
+.. automodule:: miio.fan_common
+   :members:
+   :undoc-members:
+   :show-inheritance:
