+++ docs/api/miio.fan_miot.rst
@@ -1,0 +1,7 @@
+miio.fan\_miot module
+=====================
+
+.. automodule:: miio.fan_miot
+   :members:
+   :undoc-members:
+   :show-inheritance:
